mm: full higher half support
now kernel enters `main()` with full higher half support.
identity mappings are dropped before any C code is called.
relaxed mcmodel to large to avoid headaches with the linker.

initiated documentation of the project for the sake of my sanity.

+++ kernel/kernel/main.c
@@ -9,7 +9,7 @@
 #include <lunaros/printf.h>
 #include <lunaros/tty.h>
 
-void main(uint32_t magic, uint32_t addr) {
+void main(uint64_t magic, uint64_t addr) {
    cls();
    if (magic != MULTIBOOT2_BOOTLOADER_MAGIC) {
       puts("Invalid multiboot2 magic number");
